patch 7.4.1628 Problem:    64-bit Compiler warning. Solution:   Change type of variable. (Mike Williams)

+++ src/channel.c
@@ -2099,7 +2099,7 @@ channel_part_info(channel_T *channel, dict_T *dict, char *name, int part)
 {
     chanpart_T *chanpart = &channel->ch_part[part];
     char	namebuf[20];
-    int		tail;
+    size_t	tail;
     char	*s;
 
     STRCPY(namebuf, name);
